fix: [Frontend] 세션 호출 방식 변경 및 client component 분리
